requirements/base.txt
django==5.1.2
djangorestframework==3.15.2
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
psycopg==3.2.3
drf-spectacular==0.27.2
python-decouple==3.8
psycopg2-binary==2.9.10
django-treebeard==4.7.1
whitenoise==6.8.2
drf-extra-fields==3.7.0
colorlog==6.9.0